Add files via upload
Replaced "^M" in sed scripts with "\r".
Changed how Correctable and Uncorrectable errors are stored and reported. Arris modems will present the same error count for what seems like forever. I report the error count then keep track and show "0" on the next read if it is still the same.